Fixed #55
Fixed a issue that actions would be called twice (or more than twice), or possibly cells are not even selected by user.

This is VERY ugly workaround imo, but everything else are handled in the cell class rather than in the view controller class, this implementation should keep consistency with it.