Fix FBTweak support for address sanitizer
This PR checks the address sanitizer enabled or not. If it is enabled,
every static variable will be inserted with a redzone. That will cause
the sizing information we depend on wrong. Thus, added a probe section
that contains only one fb_tweak_entry, such that we will get the exact
redzone size as well.